Actually committed stuff

+++ RPG/src/net/spencerhaney/engine/Time.java
@@ -2,7 +2,7 @@ package net.spencerhaney.engine;
 
 public class Time {
 	
-	private static final double MILLIS_IN_SECOND = 1000;
+	private static final float MILLIS_IN_SECOND = 1000;
 	
 	private static long lastTime;
 	private static long currentTime;
@@ -18,7 +18,7 @@ public class Time {
 		currentTime = System.currentTimeMillis();
 	}
 	
-	public static double getDelta(){
+	public static float getDelta(){
 		return (currentTime - lastTime) / MILLIS_IN_SECOND;
 	}
 }
